requirements.txt
--index-url http://192.168.0.113:5080/root/pypi/+simple/
#--extra-index-url https://pypi.tuna.tsinghua.edu.cn/simple
--trusted-host 192.168.0.113 pypi.tuna.tsinghua.edu.cn

pyyaml==6.0.2
requests==2.31.0
baidu-aip==2.0.0.1
pydub==0.23.1
python-dateutil==2.7.5
watchdog==0.9.0
pytz==2018.9
fire==0.1.3
tornado==6.3.3
markdown==3.0.1
semver==2.8.1
websocket==0.2.1
websocket-client==1.8.0
pypinyin==0.52.0
jieba==0.42.1
pvporcupine==3.0.3
pvrecorder==1.2.2
openai==1.40.0
apscheduler==3.10.4
asyncio==3.4.3
edge-tts==6.1.12
nest_asyncio==1.6.0
funasr_onnx==0.4.1
dbus-python==1.3.2
pyserial==3.5
pycaw==20240210
comtypes==1.4.7
